tests/requirements.txt
pytest
tqdm
scikit-learn
torchaudio==0.12.0
torchvision==0.13.0
